+++ FranAlonso/App/AppRoot.swift
@@ -1,7 +1,0 @@
-import SwiftUI
-
-struct AppRoot: View {
-    var body: some View {
-        ContentView()
-    }
-}